Feat: Observer API 호출부분 custom hooks 로 분리
